backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
google-generativeai==0.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6